DiffMICv2/requirements_kaggle.txt
# Kaggle-Compatible Requirements for DiffMICv2
# TESTED & WORKING on Kaggle (November 2025)

# Compatible versions (install with Cell 3 in KAGGLE_CODE.py)
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.4.2

# Diffusion models
peft==0.17.0
diffusers>=0.20.0
accelerate>=0.20.0

# Utilities
einops>=0.7.0
ema-pytorch>=0.1.0
easydict>=1.10

# Keep Kaggle's pre-installed versions:
# - torch (2.9.0+cu128)
# - torchvision
# - pytorch-lightning
# - torchmetrics
# - matplotlib
# - opencv-python
# - Pillow
# - pandas
# - pyyaml
